Refactor search index entry recording to dedicated type
