+++ backend/src/search/SearchIndexBuilder.ts
@@ -4,238 +4,9 @@ import { Neo4jService } from 'nest-neo4j/dist';
 import { neo4jReturnEdge, neo4jReturnNode } from '../config/commonFunctions';
 import { Node, Edge } from '../shared/entities';
 import { parseNeo4jEntityInfo } from '../schema/parseNeo4jEntityInfo';
-import { EdgeType, EntityType, NodeType } from '../shared/schema';
 import { SearchIndex } from './SearchIndex';
 import { AsyncLazy } from '../shared/utils';
-import { IndexEntry } from './IndexEntry';
-
-/**
- * Flattens the specified string array.
- */
-function flattenArray(array: string[]): string {
-  return `[${array.join(', ')}]`;
-}
-
-/**
- * Converts the specified value to string or return null if the value cannot be converted.
- * @param value The value to stringify.
- * @returns The stringified version of the value, or null of the value cannot be converted.
- */
-function convertToString(value: unknown): string | null {
-  // If the value is a string, we are done.
-  if (typeof value === 'string') {
-    return value;
-  }
-
-  // If the value is an array, convert each element and combine them via flattenArray.
-  if (Array.isArray(value)) {
-    if (value.length > 1) {
-      return flattenArray(
-        <string[]>(
-          value.map((entry) => convertToString(entry)).filter((entry) => entry)
-        )
-      );
-    }
-
-    return convertToString(value[0]);
-  }
-
-  // Try to invoke a 'toString()' function, that may be present.
-  if (typeof value === 'object' && typeof value?.toString === 'function') {
-    const result = value.toString();
-
-    if (typeof result === 'string') {
-      return result;
-    }
-  }
-
-  // The value cannot be stringified.
-  return null;
-}
-
-/**
- * Translated an edge to an index-entry and records it in an entries collection.
- * @param node The edge to translate.
- * @param entries The collection of index entries.
- * @returns The translated edge.
- */
-function recordEdge(edge: Edge, entries: IndexEntry[]): void {
-  // For an edge, we want to search in the if of the edge, in the name of the entity-type of edge
-  // and in the values of all the properties.
-  const entityIdEntry: IndexEntry = {
-    entityType: 'edge',
-    id: edge.id,
-    types: [edge.type],
-    to: edge.to,
-    from: edge.from,
-    indexKey: 'id',
-    indexValue: edge.id.toString(),
-  };
-
-  entries.push(entityIdEntry);
-
-  const entityTypeEntry: IndexEntry = {
-    entityType: 'edge',
-    id: edge.id,
-    types: [edge.type],
-    to: edge.to,
-    from: edge.from,
-    indexKey: 'type',
-    indexValue: edge.type,
-  };
-
-  entries.push(entityTypeEntry);
-
-  const keys = Object.keys(edge.properties);
-
-  // eslint-disable-next-line no-restricted-syntax
-  for (const key of keys) {
-    const value = <unknown>edge.properties[key];
-    const stringifiedValue = convertToString(value);
-
-    if (stringifiedValue) {
-      entries.push({
-        entityType: 'edge',
-        id: edge.id,
-        types: [edge.type],
-        to: edge.to,
-        from: edge.from,
-        indexKey: `properties.${key}`,
-        indexValue: stringifiedValue,
-      });
-    }
-  }
-}
-
-/**
- * Translated a node to an index-entry and records it in an entries collection.
- * @param node The node to translate.
- * @param entries The collection of index entries.
- * @returns The translated node.
- */
-function recordNode(node: Node, entries: IndexEntry[]): void {
-  // For a node, we want to search in the if of the node, in the name of the entity-type of node
-  // and in the values of all the properties.
-  const entityIdEntry: IndexEntry = {
-    entityType: 'node',
-    id: node.id,
-    types: node.types,
-    indexKey: 'id',
-    indexValue: node.id.toString(),
-  };
-
-  entries.push(entityIdEntry);
-
-  const entityTypeEntry: IndexEntry = {
-    entityType: 'node',
-    id: node.id,
-    types: node.types,
-    indexKey: 'type',
-    indexValue: flattenArray(node.types),
-  };
-
-  entries.push(entityTypeEntry);
-
-  const keys = Object.keys(node.properties);
-
-  // eslint-disable-next-line no-restricted-syntax
-  for (const key of keys) {
-    const value = <unknown>node.properties[key];
-    const stringifiedValue = convertToString(value);
-
-    if (stringifiedValue) {
-      entries.push({
-        entityType: 'node',
-        id: node.id,
-        types: node.types,
-        indexKey: `properties.${key}`,
-        indexValue: stringifiedValue,
-      });
-    }
-  }
-}
-
-/**
- * Translates an entity-type to an index-entry and records it in an entries collection.
- * @param entityType The entity-type to translates.
- * @param type The type of entity the translation affects. This is either a node-type or an edge-type.
- * @param entries The collection of index entries.
- * @returns The translated entity-type.
- */
-function recordEntityType(
-  entityType: EntityType,
-  type: 'node-type' | 'edge-type',
-  entries: IndexEntry[]
-): void {
-  // We store an entity type as index entry in the following way:
-  // The id of the entry is the name of the type,
-  // There is a single entry-property thats name is `property`and thats value is the concatenated names of
-  // all properties the entity-type contains.
-  // For example:
-  // A node of type Person consists of the properties 'name', 'born', 'address'.
-  // The index entry looks like:
-  // { id: 'Person', indexValue: '[name, born, address]' }
-
-  // For an entity-type, we want to search in the name of the entity-type
-  // and in the name of the properties, so we have to insert two entries into the index.
-  const entityTypeNameEntry = {
-    entityType: type,
-    id: entityType.name,
-    indexKey: 'id',
-    indexValue: entityType.name,
-  };
-
-  entries.push(entityTypeNameEntry);
-
-  if (entityType.properties.length > 0) {
-    const entityTypePropertiesEntry = {
-      entityType: type,
-      id: entityType.name,
-      indexKey: 'properties.properties',
-      indexValue: flattenArray(entityType.properties.map((prop) => prop.name)),
-    };
-
-    entries.push(entityTypePropertiesEntry);
-  }
-}
-
-/**
- * Translates an edge-type to an index-entry and records it in an entries collection.
- * @param edgeType The edge-type to translate.
- * @param entries The collection of index entries.
- * @returns The translated edge-type.
- */
-function recordEdgeType(edgeType: EdgeType, entries: IndexEntry[]): void {
-  const nodeEntityEntry: IndexEntry = {
-    entityType: 'edge-type',
-    id: edgeType.name,
-    indexKey: 'entity-type',
-    indexValue: 'edge',
-  };
-
-  entries.push(nodeEntityEntry);
-
-  recordEntityType(edgeType, 'edge-type', entries);
-}
-
-/**
- * Translates a node-type to an index-entry and records it in an entries collection.
- * @param nodeType The node-type to translate.
- * @param entries The collection of index entries.
- * @returns The translated node-type.
- */
-function recordNodeType(nodeType: NodeType, entries: IndexEntry[]): void {
-  const nodeEntityEntry: IndexEntry = {
-    entityType: 'node-type',
-    id: nodeType.name,
-    indexKey: 'entity-type',
-    indexValue: 'node',
-  };
-
-  entries.push(nodeEntityEntry);
-
-  recordEntityType(nodeType, 'node-type', entries);
-}
+import { SearchIndexEntryRecorder } from './SearchIndexEntryRecorder';
 
 /**
  * A search index builder that can be used to build a search index from a complete dataset.
@@ -253,7 +24,7 @@ export class SearchIndexBuilder {
   }
 
   private async buildIndexCore(): Promise<MiniSearch> {
-    const entries: IndexEntry[] = [];
+    const entryRecorder = new SearchIndexEntryRecorder();
 
     // TODO: For large databases it would be beneficial, if we read the entries in smaller batches,
     //       so we don't have to copy the entire dataset into memory (again).
@@ -272,7 +43,7 @@ export class SearchIndexBuilder {
 
     // eslint-disable-next-line no-restricted-syntax
     for (const node of nodes) {
-      recordNode(node, entries);
+      entryRecorder.recordNode(node);
     }
 
     // Get all edges from the database
@@ -290,7 +61,7 @@ export class SearchIndexBuilder {
 
     // eslint-disable-next-line no-restricted-syntax
     for (const edge of edges) {
-      recordEdge(edge, entries);
+      entryRecorder.recordEdge(edge);
     }
 
     // Get all node types from the database
@@ -302,7 +73,7 @@ export class SearchIndexBuilder {
 
     // eslint-disable-next-line no-restricted-syntax
     for (const nodeType of nodeTypes) {
-      recordNodeType(nodeType, entries);
+      entryRecorder.recordNodeType(nodeType);
     }
 
     // Get all edge types from the database
@@ -314,7 +85,7 @@ export class SearchIndexBuilder {
 
     // eslint-disable-next-line no-restricted-syntax
     for (const edgeType of edgeTypes) {
-      recordEdgeType(edgeType, entries);
+      entryRecorder.recordEdgeType(edgeType);
     }
 
     // Build the index
@@ -334,7 +105,7 @@ export class SearchIndexBuilder {
       fields,
       storeFields,
     });
-    index.addAll(entries);
+    index.addAll(entryRecorder.recordedEntries);
     return index;
   }
 }
