+++ backend/src/utils/flattenArray.ts
@@ -1,0 +1,6 @@
+/**
+ * Flattens the specified string array.
+ */
+export function flattenArray(array: string[]): string {
+  return `[${array.join(', ')}]`;
+}
